finished python communications, I have also test to check if it works. Everything works fine
